Changed data type faces array

+++ surfaceanisotropycalculator/CythonFunctions.pyx
@@ -18,8 +18,8 @@ cimport numpy as np
 from libc.math cimport sqrt
 
 
-cpdef int getinternalangles(int[:,:] f, float[:,:] v, float[:,:] result):
-    cdef int i, lenf, v0, v1, v2
+cpdef int getinternalangles(long long int[:,:] f, float[:,:] v, float[:,:] result):
+    cdef long long int i, lenf, v0, v1, v2
     cdef float[:] p0, p1, p2
     cdef float dxa, dya, dza, dxb, dyb, dzb, dxc, dyc, dzc, a1, a2, a3, temp
     lenf = f.shape[0]
@@ -54,8 +54,8 @@ cpdef int getinternalangles(int[:,:] f, float[:,:] v, float[:,:] result):
     return 1
         
         
-cpdef int getnormals(int[:,:] f, float[:,:] v, float[:,:] result):
-    cdef int i, lenf, v0, v1, v2
+cpdef int getnormals(long long int[:,:] f, float[:,:] v, float[:,:] result):
+    cdef long long int i, lenf, v0, v1, v2
     cdef float[:] p0, p1, p2, pn
     cdef float xa, ya, za, xb, yb, zb
     p0 = np.zeros(3, dtype=np.float32)
